getServerId+

+++ wf-central/src/main/java/org/igov/service/controller/SubjectController.java
@@ -505,9 +505,9 @@ public class SubjectController {
             + "\n```\n")
     @RequestMapping(value = "/getServer", method = RequestMethod.GET)
     public @ResponseBody
-    ResponseEntity getService(@ApiParam(value = "nID сервера", required = true) @RequestParam(value = "nID") Long nID) throws RecordNotFoundException {
+    ResponseEntity getService(@ApiParam(value = "nID сервера", required = true) @RequestParam(value = "nID") Integer nID) throws RecordNotFoundException {
         nID = generalConfig.getServerId(nID);
-        Optional<Server> serverOpt = serverDao.findById(nID);
+        Optional<Server> serverOpt = serverDao.findById(Long.valueOf(nID));
         if (!serverOpt.isPresent()) {
             throw new RecordNotFoundException();
         }
